+++ api-common/src/main/java/com/api/common/dto/GeoIpDTO.java
@@ -1,0 +1,20 @@
+package com.api.common.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class GeoIpDTO {
+
+    private String country;
+
+//    private String city;
+//
+//    private Double latitude;
+//
+//    private Double longitude;
+
+    private String ip;
+
+}
