+++ sink-server/pom.xml
@@ -51,5 +51,10 @@
             <scope>test</scope>
         </dependency>
 
+        <dependency>
+            <groupId>com.maxmind.geoip2</groupId>
+            <artifactId>geoip2</artifactId>
+            <version>2.16.0</version> <!-- Replace with the latest version -->
+        </dependency>
     </dependencies>
 </project>
