+++ sink-server/src/main/java/com/report/sink/handler/AppEventsHandler.java
@@ -1,0 +1,8 @@
+package com.report.sink.handler;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class AppEventsHandler {
+
+}
