+++ sink-server/src/main/java/com/report/sink/handler/FailEventLogHandler.java
@@ -1,6 +1,6 @@
 package com.report.sink.handler;
 
-import com.report.sink.helper.RedisHelper;
+import com.report.sink.properties.DataSourceProperty;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Resource;
@@ -9,13 +9,10 @@ import javax.annotation.Resource;
  * @author heqin
  */
 @Component
-public class SinkMainHandler {
+public class FailEventLogHandler {
 
     @Resource
-    private RedisHelper redisHelper;
+    private DataSourceProperty dataSourceProperty;
 
 
-    public void main() {
-
-    }
 }
