+++ sink-server/src/main/resources/application.yml
@@ -10,3 +10,7 @@ spring:
 
 mybatis-plus:
   mapper-locations: classpath*:mapper/*.xml
+
+# todo: 路径
+geo-ip:
+  file-path: /Users/heqin/workspace/java/user-behaviour-analysis/sink-server/src/main/resources/Geo-IP/Country.mmdb
